% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot}
\alias{snapshot}
\title{Transforme un url en un article pour rfrance}
\usage{
snapshot(url, title = NULL, description = NULL, site_name = NULL,
  published = NULL, twitter = NULL)
}
\description{
- extraction de quelques informations (titre, description, ...)
- webshot
- creation
}
